making changes to community and play date
Added play date to app.js

Added max_dog.png and svg to assets

Added Screen Wrapper to keep navbar at top when scrolling

Added "Play Date" section that includes text that when clicked, expands to show full message. When clicked again the text contracts.

To do:

Fix the way "Network", "Play Date", and "Chatboard" are displayed so that it is a seamless transition instead of reloading a new page.

Add a condition to prevent unnecessary re-rendering of a page if for example, you're on the network page, and you click the network tab again, the app wouldn't reload the page since you're already on it.

Build out further content for "Play Date" page.